IN PROGRESS - issue JAMESII-434: ComputationTaskStopPolicy is a generic type and needs to be handled accordingly https://altair.informatik.uni-rostock.de/jira/browse/JAMESII-434
Made ComputationTaskStopPolicyFactory generic, and adapted all concrete stop policy factories accordingly.
This also meant to change the signatures of the corresponding getters/setters in BaseExperiment, so that all code that configures a StopPolicyFactory via a ParameterizedFactory (which is the recommended approach) needs to be adapted by specific the generic parameter of the ComputationTaskStopPolicyFactory (a wildcard is sufficient).
Finally, the CompositeCompTaskStopPolicy had to be changed accordingly as well, and is now generic in the kind of computation task (IComputationTask) that is handled.